HUB_504: streamline and clean spec abTest variant
